requirements.txt
alabaster==0.7.8
amqp==2.5.2
aniso8601==9.0.1
anyjson==0.3.3
apturl==0.5.2
asgiref==3.6.0
asttokens==2.0.5
Babel==2.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
billiard==3.6.1.0
black==22.3.0
blinker==1.4
bootstrap4==0.1.0
Brlapi==0.7.0
celery==4.2.1
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
colorama==0.4.3
command-not-found==0.3
crispy-bootstrap5==0.7
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==4.1.5
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==22.2
django-bootstrap5==22.2
django-cpf-cnpj==1.0.0
django-crispy-forms==1.13.0
django-extensions==3.1.5
django-filter==22.1
django-heroku==0.3.1
django-rest-framework==0.1.0
django-tables2==2.4.1
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
ephem==3.7.7.1
executing==0.8.3
fasteners==0.14.1
feedparser==5.2.1
flake8==4.0.1
future==0.18.2
gitsome==0.8.0
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
imagesize==1.2.0
ipython==8.3.0
isort==5.10.1
jedi==0.18.1
Jinja2==2.10.1
keyring==18.0.1
kombu==4.6.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib-inline==0.1.3
mccabe==0.6.1
monotonic==1.5
mypy-extensions==0.4.3
netifaces==0.10.4
numpy==1.22.3
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.0
ply==3.11
prettytable==2.2.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.6.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.16.2
pycodestyle==2.8.0
pycups==1.9.73
pydantic==1.10.4
pyflakes==2.4.0
Pygments==2.12.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.6
python-dotenv==0.19.2
python-memcached==1.58
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2021.10.21
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
SecretStorage==2.3.1
sentry-sdk==1.5.2
setproctitle==1.1.10
simplejson==3.16.0
six==1.16.0
soupsieve==2.3.2.post1
Sphinx==1.8.5
sqlparse==0.4.2
stack-data==0.2.0
systemd-python==234
text-unidecode==1.3
tomli==1.2.1
traitlets==5.1.1
typing-extensions==4.4.0
tzdata==2022.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
validate-docbr==1.10.0
vine==1.3.0
wadllib==1.3.3
wcwidth==0.2.5
whitenoise==6.2.0
xkit==0.0.0
xonsh==0.9.13
